fix: _copyResolve1ToResolve2WebGpu

+++ src/foundation/renderer/RenderPass.ts
@@ -363,19 +363,25 @@ export class RenderPass extends RnObject {
       ) {
         continue;
       }
-      // webGpuResourceRepository.copyTextureData(
-      //   this.__resolveFrameBuffer.colorAttachments[i]._textureResourceUid,
-      //   this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid
-      // );
 
-      webGpuResourceRepository.deleteTexture(
-        this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid
-      );
-
-      this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid =
-        webGpuResourceRepository.duplicateTextureAsMipmapped(
-          this.__resolveFrameBuffer.colorAttachments[i]._textureResourceUid
+      if (
+        webGpuResourceRepository.isMippmappedTexture(
+          this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid
+        )
+      ) {
+        webGpuResourceRepository.copyTextureData(
+          this.__resolveFrameBuffer.colorAttachments[i]._textureResourceUid,
+          this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid
         );
+      } else {
+        webGpuResourceRepository.deleteTexture(
+          this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid
+        );
+        this.__resolveFrameBuffer2.colorAttachments[i]._textureResourceUid =
+          webGpuResourceRepository.duplicateTextureAsMipmapped(
+            this.__resolveFrameBuffer.colorAttachments[i]._textureResourceUid
+          );
+      }
     }
   }
